Organized components and containers. Organized components and containers by the policy. Some are changed from containers to components, removing the state.
