+++ components/EaglesEyeBoard.tsx
