+++ components/N01Board.tsx
